Add option to differentiate between dialog and focus container
While setting up a mobile menu that functions as a dialog, as described
in https://plousia.com/blog/how-create-accessible-mobile-menu, I tried
using a11y-dialog but found it impractical to stretch around this
particular edge case as-is.

These changes add the ability to instantiate a new dialog with an
options object as a second argument, inside which the key
`focusContainer` can be set to the desired element.

With the following markup and instantiation:

<nav id="focus-container">
  <button id="menu-toggle">Menu</button>
  <div id="menu-dialog">
    ...
  </div>
</nav>

let focusContainer = document.getElementById('focus-container')
let menuDialog = document.getElementById('menu-dialog')
let options = {
  focusContainer: focusContainer
}
let navMenu = new A11yDialog(menuDialog, options)

...the navigation menu's show/hide button can remain visible at all
times, and remain included in the dialog's focusable elements so that it
doesn't get excluded by tab trapping.

This pattern really probably shouldn't be used for anything else,
though.